+++ lib/activerecord-pg-format-db-structure/formatter.rb
@@ -6,14 +6,16 @@ require_relative "../activerecord-pg-format-db-structure"
 module ActiveRecordPgFormatDbStructure
   # Formats & normalizes in place the given SQL string
   class Formatter
-    attr_reader :transforms, :deparser
+    attr_reader :transforms, :deparser, :statement_appender
 
     def initialize(
       transforms: DEFAULT_TRANSFORMS,
-      deparser: DEFAULT_DEPARSER
+      deparser: DEFAULT_DEPARSER,
+      statement_appender: DEFAULT_STATEMENT_APPENDER
     )
       @transforms = transforms
       @deparser = deparser
+      @statement_appender = statement_appender
     end
 
     def format(source)
@@ -23,9 +25,17 @@ module ActiveRecordPgFormatDbStructure
         transform.new(raw_statements).transform!
       end
 
-      raw_statements.map do |raw_statement|
-        deparser.new(source).deparse_raw_statement(raw_statement)
-      end.compact.join
+      appender = statement_appender.new
+      raw_statements.each do |raw_statement|
+        statement = deparser.new(source).deparse_raw_statement(raw_statement)
+        appender.append_statement!(
+          statement,
+          statement_kind: PgQuery::Node.inner_class_to_name(
+            raw_statement.stmt.inner.class
+          )
+        )
+      end
+      appender.output
     end
   end
 end