+++ spec/activerecord-pg-format-db-structure/transforms/inline_constraints_spec.rb
@@ -50,10 +50,7 @@ RSpec.describe ActiveRecordPgFormatDbStructure::Transforms::InlineConstraints do
         ('20250124155339');
       SQL
 
-      expect(formatter.format(source)).to eq(<<~SQL.chomp)
-
-
-
+      expect(formatter.format(source)).to eq(<<~SQL)
         -- Name: comments; Type: TABLE;
 
         CREATE TABLE public.comments (