+++ spec/activerecord-pg-format-db-structure/transforms/move_indices_after_create_table_spec.rb
@@ -39,10 +39,7 @@ RSpec.describe ActiveRecordPgFormatDbStructure::Transforms::MoveIndicesAfterCrea
         ('20250124155339');
       SQL
 
-      expect(formatter.format(source)).to eq(<<~SQL.chomp)
-
-
-
+      expect(formatter.format(source)).to eq(<<~SQL)
         -- Name: comments; Type: TABLE;
 
         CREATE TABLE public.comments (